+++ test/test.js
@@ -1,5 +1,5 @@
 // i18next, v1.7.7
-// Copyright (c)2014 Jan Mühlemann (jamuhl).
+// Copyright (c)2015 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
 describe('i18next', function() {
@@ -567,13 +567,13 @@ describe('i18next', function() {
               var spy; 
       
               beforeEach(function(done) {
-                spy = sinon.spy(i18n.options, 'missingKeyHandler');
                 i18n.init(i18n.functions.extend(opts, { 
                   fallbackNS: ['ns.fallback1', 'ns.fallback2'], 
                   resStore: resStore,
                   sendMissing: true, /* must be changed to saveMissing */
                   ns: { namespaces: ['ns.common', 'ns.special', 'ns.fallback'], defaultNs: 'ns.special'} } ),
                   function(t) { 
+                    spy = sinon.spy(i18n.options, 'missingKeyHandler');
                     t('ns.common:notExisting');
                     done(); 
                   });