+++ src/i18next.functions.js
@@ -47,6 +47,15 @@ function hasResourceBundle(lng, ns) {
     return hasValues;
 }
 
+function getResourceBundle(lng, ns) {
+    if (typeof ns !== 'string') {
+        ns = o.ns.defaultNs;
+    }
+
+    resStore[lng] = resStore[lng] || {};
+    return f.extend({}, resStore[lng][ns]);
+}
+
 function removeResourceBundle(lng, ns) {
     if (typeof ns !== 'string') {
         ns = o.ns.defaultNs;