+++ releasenotes.md
@@ -3,6 +3,7 @@
 - expose applyReplacement on api
 - save resources to localStorage when useLocaleStore is true
 - add support on key is a number
+- added getResourceBundle to API
 
 ### 1.7.7
 - fixes issue with stack overflow on t(lng, count)